+++ Minigin/NullSoundService.cpp
@@ -1,6 +1,6 @@
 #include "NullSoundService.h"
 
-void dae::NullSoundService::Play(const std::string& /*soundName*/, const float /*volume*/)
+void dae::NullSoundService::Play(const std::string& /*soundName*/, const int /*volume*/)
 {
 }
