+++ Minigin/SoundSystem.h
@@ -14,7 +14,7 @@ namespace dae {
 		SoundSystem();
 		virtual ~SoundSystem() override;
 
-		virtual void Play(const std::string& soundName, const float volume) override;
+		virtual void Play(const std::string& soundName, const int volume) override;
 		virtual void Stop() override;
 	};
 }
